pkg-generic: Ensure package targets are PHONY
pkg-generic defines all the package targets for every package whether it
is selected or not in the configuration, but it only defines those
targets as PHONY if the package is selected in the configuration.

These two should go together so move the PHONY definition to where the
targets are defined.

This can make a difference for host packages that do not have any
configuration settings - the rules are used but not defined as PHONY.